Rename `data-science` plugin to `analytics`
Also removes the enabled flag
