+++ x-pack/plugin/core/src/main/java/org/elasticsearch/license/XPackLicenseState.java
@@ -70,7 +70,7 @@ public class XPackLicenseState {
             "Creating and Starting rollup jobs will no longer be allowed.",
             "Stopping/Deleting existing jobs, RollupCaps API and RollupSearch continue to function."
         });
-        messages.put(XPackField.DATA_SCIENCE, new String[] {
+        messages.put(XPackField.ANALYTICS, new String[] {
             "Aggregations provided by Data Science plugin are no longer usable."
         });
         EXPIRATION_MESSAGES = Collections.unmodifiableMap(messages);
