+++ x-pack/plugin/src/test/resources/rest-api-spec/test/analytics/cumulative_cardinality.yml
