+++ x-pack/plugin/src/test/resources/rest-api-spec/test/xpack/10_basic.yml
@@ -25,4 +25,3 @@
     - contains:  { nodes.$master.modules: { name: x-pack-security } }
     - contains:  { nodes.$master.modules: { name: x-pack-sql } }
     - contains:  { nodes.$master.modules: { name: x-pack-watcher } }
-    - contains:  { nodes.$master.modules: { name: x-pack-data-science } }
